feat(button): advanced customization
